updated to latest version, from sam.
requires bertini to be in home folder of user.  

has some code in it to run on cray xtm, but not fully functional yet.

thinking about ways to merge polysaurus and mystep2, but not there yet.  also want to eliminate the shell script calls, and move to a single system call (which I CAN do, have tested in separate program, but not implemented in polysaurus yet.

dan brake may 3 2011