this is a functional update.  modified menu systems to stem from the ProgSettings class, and propagated this class to all related functions.
we are ready to implement path failure analysis into the program with this revision.  menu skeleton is initiated, and have code which parses output files for path failures.  next is to do something about the failures other than simply notice them.

note: preferences.* is not deprecated, though not removed from the repo yet in case i find a good reason to reuse the file.  probably will be removed eventually.