updated step2 to using MPI_Send / Recv versus MPI_Bcast.  The latter has some max buffer associated with what can be sent over the wires -- or at least what i can tell -- men.
also changed several memory allocation calls from:

datatype objectname[VARIABLE_LENGTH];   to

datatype* objectname = new datatype[VARIABLE_LENGTH];

Compiler didn't catch this error due to VARIABLE_LENGTH being passed through MPI and compiler believing the user knew to pass a constant int as opposed to a variable int.

included appropriate delete[] lines for clearing memory.

finally, editted preferences for step1 to remove USERHOMOTOPY: 0 from automatically being written in the step1 config (as that is standard default anyway).  this allows for a step1 userhomotopy run, in which case, preferences was editted to test for this scenario and instead of writing 'variable_group' to the input file, we write 'variable' instead.

-- TO DO --
if USERHOMOTOPY = 1 for step1 settings, provide a location for the START FILE!

tested on sam and niemerg-mac.