apparently didn't highlight all the files for the commit (next time i will choose the folder)
updated bcasts to send / recv.  also changed instantiation of some arrays from non-variable size to variable size.

also allowed for a userhomotopy: 1 option in step1 and have editted the files accordingly to deal with this option (i.e. test for this option being 1 and writing variable and not variable_group)