gather_data_paramotopy work for failed_paths and overdetermined systems
Previously, the number of variables was assumed to be the number of equations.  Well, Priya sent me a bug where she was solving an overdetermined system and the gather function in Matlab wasn't working.  So this fixes that issue.

I also
* made the parser work for `failed_paths`, cuz it was broken.
* filtered out the files from gathering inside of paramotopy, cuz they're not useful for much else than that inside-paramotopy step
* worked on how the info files saves constants, since Priya is using expressions, not just numbers for her constants.
* cleaned up some variable names
* added a bunch of comments

i'm still not thrilled with this code, but it's less bad than it was a decade ago.  and it seems to work for these scenarios.