some cleanup on plotting functions
in particular, plotting >3 variables wasn't working correctly.  no matter your variable choices, the results were the same.  ugh.

I used the new default argument thingy via the `arguments` block which was introduced in 2019 as part of this.  finally, the language provides default argument values.  i was so sick of parsing that stuff myself.