fixed matlab gather for custom samplings, one-param runs. also, fixed solutions struct
the struct was pre-allocating as if every parameter point had the full number of solutions.  but sometimes they don't, when gathering real solutions.  so this fix makes the solutions actually be sharp, without stupid all-zero solutions as placeholders for solutions that don't actually exist.

for fixing the code for gathering custom samplings, i just had to adjust the part where the `indices` variable gets made to just be of length 1, instead of the number of parameters.

for fixing one-parameter samplings, i had to conditionally not call the `ind2sub` matlab function, which produces an error if you're converting one index to just one subscript.